fix(browser): suppress immediate-mode pending-decode warnings in useAudioTracks
In immediate mode every decode completion rebuilds all tracks from
configs; configs still awaiting their buffer (no duration/waveformData
yet) warned "Cannot create track" on every pass — quadratic console
noise for an expected transient state (12-track stem demo: dozens of
warnings). buildTrackFromConfig is extracted to its own tone-free module
(unit-testable in plain vitest) with a suppressMissingDurationWarning
option; the immediate derive pass sets it for configs whose decode is
still pending. Terminal cases (nothing left to load) still warn, and
non-immediate mode is unchanged. Live-verified: stem-tracks (12 tracks,
immediate) now logs zero such warnings.
