+++ packages/browser/src/hooks/useAudioTracks.ts
@@ -1,8 +1,6 @@
 import { useState, useEffect, useRef, useMemo } from 'react';
 import {
   ClipTrack,
-  createTrack,
-  createClipFromSeconds,
   type Fade,
   type TrackEffectsFunction,
   type WaveformDataObject,
@@ -11,6 +9,7 @@ import {
   type ColorMapValue,
 } from '@waveform-playlist/core';
 import * as Tone from 'tone';
+import { buildTrackFromConfig } from './buildTrackFromConfig';
 
 /**
  * Configuration for a single audio track to load
@@ -70,82 +69,6 @@ export interface UseAudioTracksOptions {
   immediate?: boolean;
   /** @deprecated Use `immediate` instead. */
   progressive?: boolean;
-}
-
-/** Build a ClipTrack from config + optional audioBuffer, preserving stable IDs. */
-function buildTrackFromConfig(
-  config: AudioTrackConfig,
-  index: number,
-  audioBuffer: AudioBuffer | undefined,
-  stableIds: Map<number, { trackId: string; clipId: string }>,
-  contextSampleRate: number = 48000
-): ClipTrack | null {
-  const buffer = audioBuffer ?? config.audioBuffer;
-
-  // Determine if we have enough info to create the track
-  // Prefer buffer/waveformData sample rate; fall back to the AudioContext's rate
-  const sampleRate = buffer?.sampleRate ?? config.waveformData?.sample_rate ?? contextSampleRate;
-  const sourceDuration =
-    buffer?.duration ??
-    config.waveformData?.duration ??
-    (config.duration != null ? config.duration + (config.offset ?? 0) : undefined);
-
-  if (sourceDuration === undefined) {
-    console.warn(
-      `[waveform-playlist] Track ${index + 1} ("${config.name ?? 'unnamed'}"): ` +
-        `Cannot create track — provide duration, audioBuffer, or waveformData with duration.`
-    );
-    return null;
-  }
-
-  const clip = createClipFromSeconds({
-    audioBuffer: buffer,
-    sampleRate,
-    sourceDuration,
-    startTime: config.startTime ?? 0,
-    duration: config.duration ?? sourceDuration,
-    offset: config.offset ?? 0,
-    name: config.name || `Track ${index + 1}`,
-    fadeIn: config.fadeIn,
-    fadeOut: config.fadeOut,
-    waveformData: config.waveformData,
-  });
-
-  // Validate clip values
-  if (isNaN(clip.startSample) || isNaN(clip.durationSamples) || isNaN(clip.offsetSamples)) {
-    console.error(
-      `[waveform-playlist] Invalid clip values for track ${index + 1} ("${config.name ?? 'unnamed'}"): ` +
-        `startSample=${clip.startSample}, durationSamples=${clip.durationSamples}, offsetSamples=${clip.offsetSamples}`
-    );
-    return null;
-  }
-
-  const track: ClipTrack = {
-    ...createTrack({
-      name: config.name || `Track ${index + 1}`,
-      clips: [clip],
-      muted: config.muted ?? false,
-      soloed: config.soloed ?? false,
-      volume: config.volume ?? 1.0,
-      pan: config.pan ?? 0,
-      color: config.color,
-    }),
-    effects: config.effects,
-    renderMode: config.renderMode,
-    spectrogramConfig: config.spectrogramConfig,
-    spectrogramColorMap: config.spectrogramColorMap,
-  };
-
-  // Preserve stable IDs across rebuilds so React doesn't unmount/remount tracks
-  const existingIds = stableIds.get(index);
-  if (existingIds) {
-    track.id = existingIds.trackId;
-    track.clips[0] = { ...track.clips[0], id: existingIds.clipId };
-  } else {
-    stableIds.set(index, { trackId: track.id, clipId: track.clips[0].id });
-  }
-
-  return track;
 }
 
 /**
@@ -219,12 +142,22 @@ export function useAudioTracks(configs: AudioTrackConfig[], options: UseAudioTra
 
     const result: ClipTrack[] = [];
     for (let i = 0; i < configs.length; i++) {
+      // A config whose decode hasn't arrived yet is an EXPECTED transient
+      // state in immediate mode — it builds on a later pass once the buffer
+      // lands. Suppress the missing-duration warning for those so each
+      // rebuild pass doesn't warn about every still-pending track (the
+      // count grows quadratically with track count otherwise). Configs with
+      // nothing left to load (no src, no buffer coming) keep the warning —
+      // for them, missing duration is a real misconfiguration.
+      const pendingDecode =
+        !loadedBuffers.has(i) && !configs[i].audioBuffer && configs[i].src != null;
       const track = buildTrackFromConfig(
         configs[i],
         i,
         loadedBuffers.get(i),
         stableIdsRef.current,
-        contextSampleRateRef.current
+        contextSampleRateRef.current,
+        { suppressMissingDurationWarning: pendingDecode }
       );
       if (track) result.push(track);
     }
